+++ .github/release-drafter.yml
@@ -20,6 +20,7 @@ categories:
     labels:
       - gloas
       - eip7732
+      - eip7805
       - eip7928
 
   # Features
@@ -27,8 +28,6 @@ categories:
     label: eip6110
   - title: EIP-7441
     label: eip7441
-  - title: EIP-7805
-    label: eip7805
   - title: EIP-8025
     label: eip8025
