+++ .github/workflows/checks.yml
@@ -96,7 +96,6 @@ jobs:
           - electra
           - fulu
           - gloas
-          - eip7805
           - eip7928
     steps:
       - name: Checkout repository
