+++ configs/mainnet.yaml
@@ -61,9 +61,6 @@ GLOAS_FORK_EPOCH: 18446744073709551615
 # EIP7441
 EIP7441_FORK_VERSION: 0x08000000  # temporary stub
 EIP7441_FORK_EPOCH: 18446744073709551615
-# EIP7805
-EIP7805_FORK_VERSION: 0x0a000000  # temporary stub
-EIP7805_FORK_EPOCH: 18446744073709551615
 # EIP7928
 EIP7928_FORK_VERSION: 0x0b000000  # temporary stub
 EIP7928_FORK_EPOCH: 18446744073709551615
@@ -106,8 +103,6 @@ SYNC_MESSAGE_DUE_BPS_GLOAS: 2500
 CONTRIBUTION_DUE_BPS_GLOAS: 5000
 # 7500 basis points, 75% of SLOT_DURATION_MS
 PAYLOAD_ATTESTATION_DUE_BPS: 7500
-
-# EIP7805
 # 7500 basis points, 75% of SLOT_DURATION_MS
 VIEW_FREEZE_CUTOFF_BPS: 7500
 # 6667 basis points, ~67% of SLOT_DURATION_MS
@@ -222,18 +217,16 @@ MIN_EPOCHS_FOR_DATA_COLUMN_SIDECARS_REQUESTS: 4096
 # Gloas
 # 2**7 (= 128) payloads
 MAX_REQUEST_PAYLOADS: 128
+# 2**4 (= 16) inclusion lists
+MAX_REQUEST_INCLUSION_LIST: 16
+# 2**13 (= 8,192) bytes
+MAX_BYTES_PER_INCLUSION_LIST: 8192
 
 # EIP7441
 # 2**8 (= 256) epochs
 EPOCHS_PER_SHUFFLING_PHASE: 256
 # 2**1 (= 2) epochs
 PROPOSER_SELECTION_GAP: 2
-
-# EIP7805
-# 2**4 (= 16) inclusion lists
-MAX_REQUEST_INCLUSION_LIST: 16
-# 2**13 (= 8,192) bytes
-MAX_BYTES_PER_INCLUSION_LIST: 8192
 
 # Blob Scheduling
 # ---------------------------------------------------------------
