--- presets/minimal/eip7805.yaml
@@ -1,6 +1,0 @@
-# Minimal preset - EIP7805
-
-# Inclusion list committee
-# ---------------------------------------------------------------
-# 2**4 (= 16) validators
-INCLUSION_LIST_COMMITTEE_SIZE: 16
