+++ pysetup/constants.py
@@ -9,7 +9,6 @@ FULU = "fulu"
 GLOAS = "gloas"
 EIP6800 = "eip6800"
 EIP7441 = "eip7441"
-EIP7805 = "eip7805"
 EIP7928 = "eip7928"
