+++ pysetup/md_doc_paths.py
@@ -7,7 +7,6 @@ from .constants import (
     DENEB,
     EIP6800,
     EIP7441,
-    EIP7805,
     EIP7928,
     ELECTRA,
     FULU,
@@ -26,7 +25,6 @@ PREVIOUS_FORK_OF = {
     GLOAS: FULU,
     EIP6800: DENEB,
     EIP7441: CAPELLA,
-    EIP7805: FULU,
     EIP7928: FULU,
 }
