+++ pysetup/spec_builders/__init__.py
@@ -4,7 +4,6 @@ from .capella import CapellaSpecBuilder
 from .deneb import DenebSpecBuilder
 from .eip6800 import EIP6800SpecBuilder
 from .eip7441 import EIP7441SpecBuilder
-from .eip7805 import EIP7805SpecBuilder
 from .eip7928 import EIP7928SpecBuilder
 from .electra import ElectraSpecBuilder
 from .fulu import FuluSpecBuilder
@@ -24,7 +23,6 @@ spec_builders = {
         GloasSpecBuilder,
         EIP6800SpecBuilder,
         EIP7441SpecBuilder,
-        EIP7805SpecBuilder,
         EIP7928SpecBuilder,
     )
 }
