--- pysetup/spec_builders/eip7805.py
@@ -1,67 +1,0 @@
-from ..constants import EIP7805
-from .base import BaseSpecBuilder
-
-
-class EIP7805SpecBuilder(BaseSpecBuilder):
-    fork: str = EIP7805
-
-    @classmethod
-    def imports(cls, preset_name: str):
-        return f"""
-from eth2spec.fulu import {preset_name} as fulu
-"""
-
-    @classmethod
-    def sundry_functions(cls) -> str:
-        return """
-def cached_or_new_inclusion_list_store() -> InclusionListStore:
-    # pylint: disable=unused-argument
-    return InclusionListStore()
-"""
-
-    @classmethod
-    def execution_engine_cls(cls) -> str:
-        return """
-class NoopExecutionEngine(ExecutionEngine):
-
-    def notify_new_payload(self: ExecutionEngine,
-                           execution_payload: ExecutionPayload,
-                           parent_beacon_block_root: Root,
-                           execution_requests_list: Sequence[bytes]) -> bool:
-        return True
-
-    def notify_forkchoice_updated(self: ExecutionEngine,
-                                  head_block_hash: Hash32,
-                                  safe_block_hash: Hash32,
-                                  finalized_block_hash: Hash32,
-                                  payload_attributes: Optional[PayloadAttributes]) -> Optional[PayloadId]:
-        pass
-
-    def get_payload(self: ExecutionEngine, payload_id: PayloadId) -> GetPayloadResponse:
-        # pylint: disable=unused-argument
-        raise NotImplementedError("no default block production")
-
-    def is_valid_block_hash(self: ExecutionEngine,
-                            execution_payload: ExecutionPayload,
-                            parent_beacon_block_root: Root,
-                            execution_requests_list: Sequence[bytes]) -> bool:
-        return True
-
-    def is_valid_versioned_hashes(self: ExecutionEngine, new_payload_request: NewPayloadRequest) -> bool:
-        return True
-
-    def verify_and_notify_new_payload(self: ExecutionEngine,
-                                      new_payload_request: NewPayloadRequest) -> bool:
-        return True
-
-    def get_inclusion_list(self: ExecutionEngine) -> GetInclusionListResponse:
-        # pylint: disable=unused-argument
-        raise NotImplementedError("no default inclusion list production")
-
-    def is_inclusion_list_satisfied(self: ExecutionEngine,
-                                    execution_payload: ExecutionPayload,
-                                    inclusion_list_transactions: Sequence[Transaction]) -> bool:
-        return True
-
-
-EXECUTION_ENGINE = NoopExecutionEngine()"""
