+++ specs/gloas/inclusion-list.md
@@ -1,4 +1,4 @@
-# EIP-7805 -- Inclusion List
+# Gloas -- Inclusion List
 
 <!-- mdformat-toc start --slug=github --no-anchors --maxlevel=6 --minlevel=2 -->
 
@@ -10,6 +10,8 @@
   - [New `get_inclusion_list_store`](#new-get_inclusion_list_store)
   - [New `process_inclusion_list`](#new-process_inclusion_list)
   - [New `get_inclusion_list_transactions`](#new-get_inclusion_list_transactions)
+  - [New `get_inclusion_list_bits`](#new-get_inclusion_list_bits)
+  - [New `is_inclusion_list_bits_inclusive`](#new-is_inclusion_list_bits_inclusive)
 
 <!-- mdformat-toc end -->
 
@@ -53,7 +55,7 @@ def get_inclusion_list_store() -> InclusionListStore:
 
 ```python
 def process_inclusion_list(
-    store: InclusionListStore, inclusion_list: InclusionList, is_before_view_freeze_deadline: bool
+    store: InclusionListStore, inclusion_list: InclusionList, is_before_view_freeze_cutoff: bool
 ) -> None:
     key = (inclusion_list.slot, inclusion_list.inclusion_list_committee_root)
 
@@ -72,8 +74,8 @@ def process_inclusion_list(
         # Whether it was an equivocation or not, we have processed this `inclusion_list`.
         return
 
-    # Only store `inclusion_list` if it arrived before the view freeze deadline.
-    if is_before_view_freeze_deadline:
+    # Only store `inclusion_list` if it arrived before the view freeze cutoff.
+    if is_before_view_freeze_cutoff:
         store.inclusion_lists[key].add(inclusion_list)
 ```
 
@@ -96,9 +98,50 @@ def get_inclusion_list_transactions(
     inclusion_list_transactions = [
         transaction
         for inclusion_list in store.inclusion_lists[key]
+        if inclusion_list.validator_index not in store.equivocators[key]
         for transaction in inclusion_list.transactions
     ]
 
     # Deduplicate inclusion list transactions. Order does not need to be preserved.
     return list(set(inclusion_list_transactions))
+```
+
+### New `get_inclusion_list_bits`
+
+```python
+def get_inclusion_list_bits(
+    store: InclusionListStore, state: BeaconState, slot: Slot
+) -> Bitvector[INCLUSION_LIST_COMMITTEE_SIZE]:
+    inclusion_list_committee = get_inclusion_list_committee(state, slot)
+    inclusion_list_committee_root = hash_tree_root(inclusion_list_committee)
+    key = (slot, inclusion_list_committee_root)
+
+    validator_indices = [
+        inclusion_list.validator_index for inclusion_list in store.inclusion_lists[key]
+    ]
+
+    return Bitvector[INCLUSION_LIST_COMMITTEE_SIZE](
+        validator_index in validator_indices for validator_index in inclusion_list_committee
+    )
+```
+
+### New `is_inclusion_list_bits_inclusive`
+
+```python
+def is_inclusion_list_bits_inclusive(
+    store: InclusionListStore,
+    state: BeaconState,
+    slot: Slot,
+    inclusion_list_bits: Bitvector[INCLUSION_LIST_COMMITTEE_SIZE],
+) -> bool:
+    inclusion_list_store = get_inclusion_list_store()
+
+    local_inclusion_list_bits = get_inclusion_list_bits(inclusion_list_store, state, Slot(slot - 1))
+
+    return all(
+        inclusion_bit or not local_inclusion_bit
+        for inclusion_bit, local_inclusion_bit in zip(
+            inclusion_list_bits, local_inclusion_list_bits
+        )
+    )
 ```
